requirements.txt
langchain
langchain-community
transformers
sentence-transformers
faiss-cpu
PyMuPDF
